+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- Fix issue where script gbvm symbols would not always match the ones you defined
+- Fix issue where custom script gbvm symbols would not always match the ones you defined
+- Fix issue where identical custom scripts would share the same gbvm symbol
 
 ## [4.1.0] - 2024-09-02